requirements.txt
absl-py==1.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
comtypes==1.1.14
config==0.5.1
cryptography==38.0.3
DateTime==4.7
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
Django==4.1.1
dlib==19.24.0
docutils==0.19
entrypoints==0.4
executing==1.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.8.0
firebase-admin==6.0.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
flask-paginate==2022.1.8
Flask-SQLAlchemy==3.0.0
flatbuffers==2.0.7
gast==0.4.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.21.0
idna==3.4
imageio==2.21.2
imageio-ffmpeg==0.4.7
imutils==0.5.4
inflect==6.0.2
ipykernel==6.15.3
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.10.0
Keras-Preprocessing==1.1.2
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==1.0.4
mysql-connector==2.2.9
mysqlclient==2.1.1
nest-asyncio==1.5.5
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pendulum==2.1.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
proglog==0.1.10
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyGetWindow==0.0.9
Pygments==2.13.0
PyJWT==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.21.0
pyttsx3==2.90
pytz==2022.6
pytzdata==2020.1
pywin32==303
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
stack-data==0.2.0
syrics==0.0.1.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==2.0.1
termcolor2==0.0.3
tf-estimator-nightly==2.8.0.dev2021122109
tinytag==1.8.1
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.0
zope.interface==5.5.0
